RESTEasy can be excluded if using code.quarkus.io
Reads the data from code.quarkus.io or code.quarkus.redhat.com
to see which quarkus extensions are required.
This means if code.quarkus.io is used, then resteasy is not required.
Reads the Open API of the code quarkus instance to determine if
the option to prevent sample code being generated is present.
If it is, then it adds another step to the project
generation wizard to specify if sample code should be generated.

Closes #322

Signed-off-by: David Thompson <davthomp@redhat.com>